Extract functionality for find-certificate into service class
